refactor(settings): extract category sections with groups
Reallocate the 37 items into category sections without changing the
single-scroll entry (entry branching lands in P1-C3).

- sections/display_section.dart: Display 13 items in 3 groups
  (Appearance -> Terminal -> Key Overlay) still referencing the old
  settingsSection* keys until the new settingsGroup* keys arrive (P1-C4).
- sections/behavior_section.dart: 11 flat items, bundling
  wheelSendVerifiedProvider + settingsScrollSendUnverifiedNote, keeping
  the iOS-only CJK check on Theme.of(context).platform and adding the
  known-unconsumed note for enableVibration.
- sections/connection_section.dart: Image Transfer group (9) + flat
  Clear SSH Host Keys.
- sections/about_section.dart: 3 flat items.
- settings_screen.dart: single scroll assembled from the four sections
  (Display -> Behavior -> Connection -> About) with dividers.
- test helper scrollUntilFound now ensures the target is fully on-screen
  after scrollUntilVisible (lazy slivers can stop on a built-but-offscreen
  target once the new ordering shifts item positions).
